backend/cuda: move yaksuri_cuda function to its own file
Defining the function as a static in a header is hacky and unnecessary
for performance.
