+++ src/backend/cuda/pup/Makefile.mk
@@ -6,6 +6,7 @@
 AM_CPPFLAGS += -I$(top_srcdir)/src/backend/cuda/pup
 
 libyaksa_la_SOURCES += \
+	src/backend/cuda/pup/yaksuri_cuda_buftype.c \
 	src/backend/cuda/pup/yaksuri_cudai_pup_char.cu \
 	src/backend/cuda/pup/yaksuri_cudai_pup_double.cu \
 	src/backend/cuda/pup/yaksuri_cudai_pup_float.cu \
